Refactor Markdown chunker
