+++ native/src/test/java/io/ballerina/stdlib/ai/MarkdownChunkerIntegrationTest.java
@@ -1,9 +1,6 @@
 package io.ballerina.stdlib.ai;
 
 import dev.langchain4j.data.segment.TextSegment;
-import org.testng.Assert;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -12,6 +9,10 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import org.testng.Assert;
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Test;
 
 public class MarkdownChunkerIntegrationTest {
 
@@ -368,24 +369,24 @@ public class MarkdownChunkerIntegrationTest {
     public void testBreakUpChunkMetadataLinking() {
         // Create a test chunk with some metadata
         Map<String, String> originalMetadata = Map.of("type", "test", "source", "unit_test");
-        MarkdownChunker.Chunk originalChunk = new MarkdownChunker.Chunk(
+        RecursiveChunker.Chunk originalChunk = new RecursiveChunker.Chunk(
                 "This is a very long text that needs to be broken up into smaller chunks", originalMetadata);
 
         // Break up the chunk with a small max size to force multiple chunks
-        List<MarkdownChunker.Chunk> brokenChunks = MarkdownChunker.breakUpChunk(originalChunk, 10);
+        List<RecursiveChunker.Chunk> brokenChunks = RecursiveChunker.breakUpChunk(originalChunk, 10);
 
         // Verify we got multiple chunks
         Assert.assertTrue(brokenChunks.size() > 1, "Should create multiple chunks when breaking up large content");
 
         // Verify the first chunk doesn't have a prev link (it's the first)
-        MarkdownChunker.Chunk firstChunk = brokenChunks.get(0);
+        RecursiveChunker.Chunk firstChunk = brokenChunks.get(0);
         Assert.assertFalse(firstChunk.metadata().containsKey("prev"),
                 "First chunk should not have a prev link");
 
         // Verify subsequent chunks have prev links to the previous chunk
         for (int i = 1; i < brokenChunks.size(); i++) {
-            MarkdownChunker.Chunk currentChunk = brokenChunks.get(i);
-            MarkdownChunker.Chunk previousChunk = brokenChunks.get(i - 1);
+            RecursiveChunker.Chunk currentChunk = brokenChunks.get(i);
+            RecursiveChunker.Chunk previousChunk = brokenChunks.get(i - 1);
 
             Assert.assertTrue(currentChunk.metadata().containsKey("prev"),
                     "Chunk " + i + " should have a prev link");
@@ -396,7 +397,7 @@ public class MarkdownChunkerIntegrationTest {
         }
 
         // Verify original metadata is preserved in all chunks
-        for (MarkdownChunker.Chunk chunk : brokenChunks) {
+        for (RecursiveChunker.Chunk chunk : brokenChunks) {
             Assert.assertEquals(chunk.metadata().get("type"), "test",
                     "All chunks should preserve original type metadata");
             Assert.assertEquals(chunk.metadata().get("source"), "unit_test",
@@ -405,7 +406,7 @@ public class MarkdownChunkerIntegrationTest {
 
         // Verify the combined text equals the original
         String combinedText = brokenChunks.stream()
-                .map(MarkdownChunker.Chunk::piece)
+                .map(RecursiveChunker.Chunk::piece)
                 .collect(Collectors.joining());
         Assert.assertEquals(combinedText, originalChunk.piece(),
                 "Combined text from broken chunks should equal original text");
@@ -415,9 +416,9 @@ public class MarkdownChunkerIntegrationTest {
     public void testBreakUpChunkEdgeCases() {
         // Test with chunk smaller than max size
         Map<String, String> metadata = Map.of("type", "small");
-        MarkdownChunker.Chunk smallChunk = new MarkdownChunker.Chunk("Small", metadata);
+        RecursiveChunker.Chunk smallChunk = new RecursiveChunker.Chunk("Small", metadata);
 
-        List<MarkdownChunker.Chunk> result = MarkdownChunker.breakUpChunk(smallChunk, 10);
+        List<RecursiveChunker.Chunk> result = RecursiveChunker.breakUpChunk(smallChunk, 10);
 
         Assert.assertEquals(result.size(), 1, "Should return single chunk when content is smaller than max size");
         Assert.assertEquals(result.get(0).piece(), "Small", "Should preserve original content");
@@ -425,14 +426,14 @@ public class MarkdownChunkerIntegrationTest {
                 "Single chunk should not have prev link");
 
         // Test with empty chunk
-        MarkdownChunker.Chunk emptyChunk = new MarkdownChunker.Chunk("", metadata);
-        List<MarkdownChunker.Chunk> emptyResult = MarkdownChunker.breakUpChunk(emptyChunk, 10);
+        RecursiveChunker.Chunk emptyChunk = new RecursiveChunker.Chunk("", metadata);
+        List<RecursiveChunker.Chunk> emptyResult = RecursiveChunker.breakUpChunk(emptyChunk, 10);
 
         Assert.assertEquals(emptyResult.size(), 0, "Should return empty list for empty chunk");
 
         // Test with chunk exactly at max size
-        MarkdownChunker.Chunk exactChunk = new MarkdownChunker.Chunk("Exactly10!", metadata);
-        List<MarkdownChunker.Chunk> exactResult = MarkdownChunker.breakUpChunk(exactChunk, 10);
+        RecursiveChunker.Chunk exactChunk = new RecursiveChunker.Chunk("Exactly10!", metadata);
+        List<RecursiveChunker.Chunk> exactResult = RecursiveChunker.breakUpChunk(exactChunk, 10);
 
         Assert.assertEquals(exactResult.size(), 1, "Should return single chunk when content is exactly max size");
         Assert.assertEquals(exactResult.get(0).piece(), "Exactly10!", "Should preserve original content");
